Refactor : togglePasswordVisibility 로직 분리
