+++ src/app/(form-layout)/reset-password/_reset-password/components/ResetPasswordForm.tsx
@@ -3,30 +3,21 @@ import PasswordToggleButton from '@/app/(form-layout)/signup/_signup/PasswordTog
 import Button from '@/components/common/Button';
 import FormField from '@/components/common/formField';
 import axiosClient from '@/lib/axiosClient';
+import togglePasswordVisibilityLogic from '@/utils/toggle-password-visibility';
 import { validateConfirmPassword, validatePassword } from '@/utils/validators';
 import { useState } from 'react';
 
 /**
  * @todo
- * 처음 렌더링 되었을때 버튼 에러 표시 되어있는 오류 해결
+ * 처음 렌더링 되었을때 인풋이 에러 표시 되어있는 오류 해결
  */
 export default function ResetPasswordForm() {
+  const { isPasswordVisible, togglePasswordVisibility } = togglePasswordVisibilityLogic();
+
   const [formData, setFormData] = useState({
     password: '',
     passwordConfirmation: '',
   });
-
-  const [isPasswordVisible, setIsPasswordVisible] = useState({
-    password: false,
-    confirmPassword: false,
-  });
-
-  const togglePasswordVisibility = (key: 'password' | 'confirmPassword') => {
-    setIsPasswordVisible((prev) => ({
-      ...prev,
-      [key]: !prev[key],
-    }));
-  };
 
   const FormFieldArray = [
     {
